Add option to specify the word to be shown with the feedback
"
I'd like to propose a change about how the feedback is displayed:
if the user enters a keyword from the alternatives list, the feedback
table still displays the "main keyword". It would be better if the
keyword in the table was the actually entered keyword.
"

See https://github.com/otacke/h5p-essay/issues/21